requirements.txt
langchain
langchain-groq
langchain-openai
openai
openrouter
requests
python-dotenv
fpdf
fastapi
uvicorn
pytesseract
PyMuPDF
Pillow
faiss-cpu
langchain-google-genai
langchain-chroma
pypdf
